add координаты для навигации
